+++ packages/storybook-chakra-ui/src/FlexOverflow.indexer.ts
@@ -1,0 +1,3 @@
+import flexIndexer from "./Flex.indexer";
+
+export default flexIndexer;