+++ packages/web/src/index.ts
@@ -1,3 +1,0 @@
-import "./Box";
-import "./Flex";
-import "./FlexItem";