+++ packages/storybook-chakra-ui/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @moblin/storybook-chakra-ui
 
+## 1.2.0-esm.3
+
+### Minor Changes
+
+- c30fe51: big refactoring in tooling for esm modules
+
+### Patch Changes
+
+- Updated dependencies [c30fe51]
+  - @moblin/storybook-indexer@0.2.0-esm.3
+  - @moblin/chakra-ui@2.3.0-esm.3
+  - @moblin/core@0.2.0-esm.3
+
+## 1.2.0-esm.2
+
+### Patch Changes
+
+- obey node esm module imports with extensions
+
+## 1.2.0-esm.1
+
+### Patch Changes
+
+- fix exports for types
+
+## 1.2.0-esm.0
+
+### Minor Changes
+
+- export esm modules, update dependencies
+
 ## 1.1.0
 
 ### Minor Changes